requirements.txt
google-adk
python-dotenv
crawl4ai>=0.3.0
beautifulsoup4>=4.9.0
lxml>=4.9.0
requests>=2.25.0
